requirements.txt
pytest==7.1.2
mlflow==1.27.0
pandas==1.4.3
sklearn==0.0
azure-identity
azure.ai.ml
scikit-learn==1.1.1
flake8
